Deleted unused imports (random and time); added from bs4.element import ResultSet; added getLogger; added exception handler to get_duration, to handle ValueError; changed regex patter in get_youtube_id to get youtube ID for a vidoe by link; devide get_data into two separate functions: The get_data() function retrieves podcast episode data from the web page and collects the title, guest name, description, duration, YouTube URL, audio file URL, and thumbnail URL for each episode. The data is returned in the form of a set of tuples, where each tuple contains the metadata for a single podcast episode.
The parse_the_data() function is used to parse the data for a single episode, it takes a BeautifulSoup ResultSet object as input, then extracts the relevant data and returns a tuple containing the parsed information.

Added main function that calls the get_data function to retrieve all the podcast episodes from the Lex Fridman podcast webpage and passes each episode to the parse_the_data function to extract the relevant metadata for the podcast. Finally, it writes the retrieved metadata to a CSV file using the save_list_to_csv function.